rename binary => bytes
