+++ sdk/src/env.rs
@@ -42,7 +42,7 @@ pub type EnvType<V> = internal::EnvVal<Env, V>;
 pub type EnvVal = internal::EnvVal<Env, RawVal>;
 pub type EnvObj = internal::EnvVal<Env, Object>;
 
-use crate::binary::{Bytes, BytesN};
+use crate::bytes::{Bytes, BytesN};
 use crate::ContractData;
 
 /// The [Env] type provides access to the environment the contract is executing
