+++ sdk/src/lib.rs
@@ -61,7 +61,7 @@ pub use envhidden::*;
 
 mod account;
 mod bigint;
-mod binary;
+mod bytes;
 mod contract_data;
 pub mod iter;
 mod map;
@@ -69,8 +69,8 @@ mod vec;
 pub use account::Account;
 pub use bigint::BigInt;
 #[allow(deprecated)]
-pub use binary::{Binary, FixedBinary};
-pub use binary::{Bytes, BytesN};
+pub use bytes::{Binary, FixedBinary};
+pub use bytes::{Bytes, BytesN};
 pub use contract_data::ContractData;
 pub use map::Map;
 pub use vec::Vec;
